requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-crispy-forms==1.14.0
django-filter==21.1
idna==3.3
phonenumbers==8.12.51
pycryptodome==3.14.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
sw-python-utils==0.0.18
tzdata==2022.1
urllib3==1.26.9
